GameMaster/backend/utils/tracing.py
"""
Langfuse tracing configuration.
Initialize once in main.py, then all @observe() decorators work automatically.
"""
import os
from dotenv import load_dotenv


def init_tracing():
    """
    Initialize Langfuse tracing.

    Langfuse uses environment variables automatically:
    - LANGFUSE_PUBLIC_KEY
    - LANGFUSE_SECRET_KEY
    - LANGFUSE_BASE_URL

    If these are set in .env, tracing will be enabled.
    If not, the app will still work but without tracing.
    """
    load_dotenv()

    required = [
        "LANGFUSE_PUBLIC_KEY",
        "LANGFUSE_SECRET_KEY",
        "LANGFUSE_BASE_URL"
    ]

    missing = [key for key in required if not os.getenv(key)]

    if missing:
        print("⚠️  Langfuse tracing not configured")
        print(f"   Missing: {', '.join(missing)}")
        print("   App will run without tracing")
        return False
    else:
        print("✅ Langfuse tracing enabled")
        print(f"   Base URL: {os.getenv('LANGFUSE_BASE_URL')}")
        return True
